refactor: fix domains packages inside applications
