+++ application/src/main/java/com/piresvet/gateway/Vet/CreateVetGateway.java
@@ -1,0 +1,7 @@
+package com.piresvet.gateway.Vet;
+
+import com.piresvet.core.domain.Vet;
+
+public interface CreateVetGateway {
+    Vet create(Vet vet);
+}
