+++ application/src/main/java/com/piresvet/gateway/Vet/DeleteVetGateway.java
@@ -1,4 +1,4 @@
-package com.piresvet.gateway;
+package com.piresvet.gateway.Vet;
 
 import java.util.UUID;
