+++ application/src/main/java/com/piresvet/gateway/Vet/UpdateVetUseGateway.java
@@ -1,4 +1,4 @@
-package com.piresvet.gateway;
+package com.piresvet.gateway.Vet;
 
 import com.piresvet.core.domain.Vet;
