+++ application/src/main/java/com/piresvet/usecaseimpl/Vet/FindVetByCrmvUseCaseImpl.java
@@ -1,8 +1,8 @@
-package com.piresvet.usecaseimpl;
+package com.piresvet.usecaseimpl.Vet;
 
 import com.piresvet.core.domain.Vet;
 import com.piresvet.core.exception.InvalidCrmvException;
-import com.piresvet.gateway.FindVetByCrmvGateway;
+import com.piresvet.gateway.Vet.FindVetByCrmvGateway;
 import com.piresvet.usecase.Vet.FindVetByCrmvUseCase;
 
 public class FindVetByCrmvUseCaseImpl implements FindVetByCrmvUseCase {
