+++ application/src/main/java/com/piresvet/usecaseimpl/Vet/UpdateVetUseCaseImpl.java
@@ -1,8 +1,8 @@
-package com.piresvet.usecaseimpl;
+package com.piresvet.usecaseimpl.Vet;
 
 import com.piresvet.core.domain.Vet;
-import com.piresvet.gateway.FindVetByIdGateway;
-import com.piresvet.gateway.UpdateVetUseGateway;
+import com.piresvet.gateway.Vet.FindVetByIdGateway;
+import com.piresvet.gateway.Vet.UpdateVetUseGateway;
 import com.piresvet.usecase.Vet.UpdateVetUseCase;
 
 import java.util.UUID;
